修改票数数据类型为int

+++ src/main/java/com/blue/yw/repository/VoteRepository.java
@@ -10,6 +10,6 @@ import java.util.List;
 public interface VoteRepository extends JpaRepository<VoteEntity, Integer> {
     List<VoteEntity> findByState(String state);
 
-    Long countByNominationId(Integer nominationId);
+    Integer countByNominationId(Integer nominationId);
 
 }
